# fix(db/boltdb): consolidate tiny txs