--- a/pkg/db/common/boltdb/boltdb.go
+++ b/pkg/db/common/boltdb/boltdb.go
@@ -558,15 +558,15 @@ func (c *Connection) PutVulnerabilityData(root string) error {
 			return errors.Wrapf(err, "walk %s", root)
 		}
 
+		for _, root := range roots {
+			if err := putRoot(tx, root); err != nil {
+				return errors.Wrap(err, "put root")
+			}
+		}
+
 		return nil
 	}); err != nil {
 		return errors.WithStack(err)
-	}
-
-	for _, root := range roots {
-		if err := putRoot(c.conn, root); err != nil {
-			return errors.Wrap(err, "put root")
-		}
 	}
 
 	return nil
@@ -735,68 +735,46 @@ func putVulnerability(tx *bolt.Tx, data dataTypes.Data, roots map[string]dbTypes
 	return nil
 }
 
-func putRoot(conn *bolt.DB, root dbTypes.VulnerabilityRoot) error {
-	if err := conn.View(func(tx *bolt.Tx) error {
-		vb := tx.Bucket([]byte("vulnerability"))
-		if vb == nil {
-			return errors.Errorf("bucket:%q is not exists", "vulnerability")
-		}
-
-		vrb := vb.Bucket([]byte("root"))
-		if vrb == nil {
-			return nil
-		}
-
-		bs := vrb.Get([]byte(root.ID))
-		if len(bs) > 0 {
-			var r dbTypes.VulnerabilityRoot
-			if err := util.Unmarshal(bs, &r); err != nil {
-				return errors.Wrapf(err, "unmarshal %s", fmt.Sprintf("vulnerability:root:%s", r.ID))
-			}
-			for _, a := range r.Advisories {
-				if !slices.Contains(root.Advisories, a) {
-					r.Advisories = append(r.Advisories, a)
-				}
-			}
-			for _, v := range r.Vulnerabilities {
-				if !slices.Contains(root.Vulnerabilities, v) {
-					r.Vulnerabilities = append(r.Vulnerabilities, v)
-				}
-			}
-			for _, d := range r.DataSources {
-				if !slices.Contains(root.DataSources, d) {
-					root.DataSources = append(root.DataSources, d)
-				}
-			}
-		}
-		return nil
-	}); err != nil {
-		return errors.Wrapf(err, "view %s", fmt.Sprintf("vulnerability:root:%s", root.ID))
+func putRoot(tx *bolt.Tx, root dbTypes.VulnerabilityRoot) error {
+	vb := tx.Bucket([]byte("vulnerability"))
+	if vb == nil {
+		return errors.Errorf("bucket:%q is not exists", "vulnerability")
 	}
 
-	if err := conn.Update(func(tx *bolt.Tx) error {
-		vb := tx.Bucket([]byte("vulnerability"))
-		if vb == nil {
-			return errors.Errorf("bucket:%q is not exists", "vulnerability")
-		}
-
-		vrb, err := vb.CreateBucketIfNotExists([]byte("root"))
-		if err != nil {
-			return errors.Wrapf(err, "create bucket:%q if not exists", "vulnerability:root")
-		}
-
-		bs, err := util.Marshal(root)
-		if err != nil {
-			return errors.Wrap(err, "marshal root")
-		}
-
-		if err := vrb.Put([]byte(root.ID), bs); err != nil {
-			return errors.Wrapf(err, "put %s", fmt.Sprintf("vulnerability:root:%s", root.ID))
-		}
-
+	vrb := vb.Bucket([]byte("root"))
+	if vrb == nil {
 		return nil
-	}); err != nil {
-		return errors.Wrapf(err, "update %s", fmt.Sprintf("vulnerability:root:%s", root.ID))
+	}
+
+	if bs := vrb.Get([]byte(root.ID)); len(bs) > 0 {
+		var r dbTypes.VulnerabilityRoot
+		if err := util.Unmarshal(bs, &r); err != nil {
+			return errors.Wrapf(err, "unmarshal %s", fmt.Sprintf("vulnerability:root:%s", r.ID))
+		}
+		for _, a := range r.Advisories {
+			if !slices.Contains(root.Advisories, a) {
+				r.Advisories = append(r.Advisories, a)
+			}
+		}
+		for _, v := range r.Vulnerabilities {
+			if !slices.Contains(root.Vulnerabilities, v) {
+				r.Vulnerabilities = append(r.Vulnerabilities, v)
+			}
+		}
+		for _, d := range r.DataSources {
+			if !slices.Contains(root.DataSources, d) {
+				root.DataSources = append(root.DataSources, d)
+			}
+		}
+	}
+
+	bs, err := util.Marshal(root)
+	if err != nil {
+		return errors.Wrap(err, "marshal root")
+	}
+
+	if err := vrb.Put([]byte(root.ID), bs); err != nil {
+		return errors.Wrapf(err, "put %s", fmt.Sprintf("vulnerability:root:%s", root.ID))
 	}
 
 	return nil