@MaineK00n 2024-07-25
> vulnerability:root bucket not found...
> 
> ```
> $ vuls db init
> $ vuls db add ~/extracted/vuls-data-extracted-amazon
> 
> $ bbolt keys ~/.cache/vuls/vuls.db vulnerability
> advisory
> vulnerability
> ```

```diff
```diff
:100644 100644 91462cc 0000000 M	pkg/db/common/boltdb/boltdb.go

diff --git a/pkg/db/common/boltdb/boltdb.go b/pkg/db/common/boltdb/boltdb.go
index 91462cc..c900355 100644
--- a/pkg/db/common/boltdb/boltdb.go
+++ b/pkg/db/common/boltdb/boltdb.go
@@ -741,9 +741,9 @@ func putRoot(tx *bolt.Tx, root dbTypes.VulnerabilityRoot) error {
 		return errors.Errorf("bucket:%q is not exists", "vulnerability")
 	}
 
-	vrb := vb.Bucket([]byte("root"))
-	if vrb == nil {
-		return nil
+	vrb, err := vb.CreateBucketIfNotExists([]byte("root"))
+	if err != nil {
+		return errors.Wrapf(err, "create bucket:%q if not exists", "vulnerability:root")
 	}
 
 	if bs := vrb.Get([]byte(root.ID)); len(bs) > 0 {

```
```